requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
requests
google-cloud-storage
